refactor: update_state_source
