+++ crates/pop-cli/src/commands/mod.rs
@@ -142,6 +142,8 @@ impl Command {
 					test::Command::Contract(cmd) => cmd.execute(&mut Cli).await.map(|t| json!(t)),
 					#[cfg(feature = "parachain")]
 					test::Command::OnRuntimeUpgrade(cmd) => cmd.execute(&mut Cli).await.map(|t| json!(t)),
+					#[cfg(feature = "parachain")]
+					test::Command::ExecuteBlock(cmd) => cmd.execute(&mut Cli).await.map(|t| json!(t)),
 				},
 			},
 			Self::Clean(args) => match args.command {
